feat: added parameterized days to show per user

+++ menus-ionic/src/app/main/week/pages/week-page/week-page.component.ts
@@ -1,10 +1,14 @@
 import { Component } from '@angular/core';
 import { WeekMeal } from "@models/week-meal.model";
 import { ActivatedRoute, Router } from "@angular/router";
-import { AlertController } from "@ionic/angular";
+import { AlertController, PopoverController } from "@ionic/angular";
 import { WeekService } from "@services/week.service";
 import { ConfirmationAlertService } from "@services/confirmation-alert.service";
 import { getMealDayStringified, MealDay } from "@models/enums/meal-day.enum";
+import { OptionPopoverComponent } from "@components/popover/option-popover/option-popover.component";
+import { FoodAuthService } from "@services/food-auth.service";
+import { UserRestService } from "@services/user-rest.service";
+import { User } from "@models/user.model";
 
 @Component({
     selector: 'app-week-page',
@@ -34,12 +38,17 @@ export class WeekPageComponent {
     meals: WeekMeal[] = [];
     isDeleting: number = undefined;
 
+    private currentUser: User;
+
     constructor(
         private weekService: WeekService,
         private router: Router,
         private route: ActivatedRoute,
         private alertController: AlertController,
-        private confirmationService: ConfirmationAlertService
+        private confirmationService: ConfirmationAlertService,
+        private popoverController: PopoverController,
+        private authService: FoodAuthService,
+        private userService: UserRestService
     ) {
         this.meals = this.weekService.meals$.getValue();
         this.changeTodayMeal();
@@ -48,6 +57,9 @@ export class WeekPageComponent {
             this.meals = m;
             this.changeTodayMeal();
         });
+
+        this.currentUser = this.authService.user.getValue();
+        this.authService.user.subscribe(u => this.currentUser = u);
     }
 
     changeTodayMeal() {
@@ -69,10 +81,14 @@ export class WeekPageComponent {
     }
 
     getDay(mealDay: MealDay) {
-        return getMealDayStringified(mealDay);
+        return getMealDayStringified(mealDay.valueOf());
     }
 
-    clearRecette(index: number) {
+    isHidden(mealDay: MealDay) {
+        return this.authService.user.getValue().daysToShow.findIndex(m => m.valueOf() === mealDay.valueOf()) === -1;
+    }
+
+    clearRecipe(index: number) {
         this.weekService.deleteMeal(index);
     }
 
@@ -96,26 +112,7 @@ export class WeekPageComponent {
     }
 
     async deleteWeekMeal(i: number) {
-        const alert = await this.alertController.create({
-            header: 'Confirmation',
-            cssClass: 'confirmation-modal',
-            message: 'Supprimer ce repas de la semaine ?',
-            buttons: [
-                {
-                    text: 'Annuler',
-                    role: 'cancel',
-                    cssClass: 'cancel'
-                }, {
-                    cssClass: 'confirmation',
-                    text: 'Okay',
-                    handler: () => {
-                        this.weekService.deleteMeal(i);
-                    }
-                }
-            ]
-        });
-
-        await alert.present();
+        await this.confirmationService.confirm("Supprimer le repas de " + this.getDay(i) + " ?", () => this.weekService.deleteMeal(i));
     }
 
     gotToMeal(i: number) {
@@ -129,4 +126,69 @@ export class WeekPageComponent {
             this.isDeleting = undefined;
         }
     }
+
+    async showOptions(event: any) {
+        const options = [
+            { clickedResult: "DAYS", text: "Choisir les jours à afficher" }
+        ];
+
+        const popover = await this.popoverController.create({
+            component: OptionPopoverComponent,
+            event: event,
+            translucent: true,
+            componentProps: {
+                options
+            }
+        });
+        popover.present();
+
+        const { data } = await popover.onWillDismiss();
+        if (data && data.option) {
+            if (data.option === "DAYS") {
+                this.showDaysOptions();
+            }
+        }
+    }
+
+    async showDaysOptions() {
+        const user = this.authService.user.getValue();
+        if (!user.daysToShow)
+            user.daysToShow = [];
+        const allDays = Object.keys(MealDay).filter(k => typeof MealDay[k as any] === "number").map(k => MealDay[k as any]);
+        const inputs = [];
+
+        allDays.forEach(val => {
+            const intValue = parseInt(val);
+            inputs.push({
+                name: 'day',
+                type: 'checkbox',
+                label: getMealDayStringified(intValue),
+                value: val,
+                checked: user.daysToShow.findIndex(meal => meal.valueOf() === intValue) > -1
+            });
+        });
+
+        const alert = await this.alertController.create({
+            header: 'Choisir les jours à afficher',
+            inputs: inputs,
+            buttons: [
+                {
+                    text: 'Annuler',
+                    role: "cancel"
+                }, {
+                    text: 'Valider',
+                    handler: (alertData) => {
+                        user.daysToShow = alertData;
+                        this.userService.saveUser(user).subscribe(user => {
+                           if (this.authService.user.getValue().id === user.id)
+                               this.authService.user.next(user);
+                        });
+                    }
+                }
+            ]
+        });
+
+        await alert.present();
+    }
+
 }
